pyproject.toml
[project]
name = "mlops-serve-mobilenetv3-plain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]
